Seperate known unused stuff from syncdata
